// apps/planes_de_trabajo/src/app/core/models/curso.model.ts
export interface Curso {
  numIdProfesor: string;
  periodo: string;
  codAsignatura: string;
  grupo: string;
  semestre: number;
  nomAsignatura: string;
  numEstudiantes: number;
  numCreditos: number;
  horaInicio: string;
  horaFin: string;
  horasPresenciales: number;
  codPrograma: string;
  nomPrograma: string;
  centroCosto: string;
}